refactor: extract Epic Fight's screen processors out of EpicFightControlifyEntrypoint
